+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-- `update graphiql@1.5.17 to graphiql@1.6.0`
+- `update graphiql@1.5.17 to graphiql@1.7.0`
+- `update @tarantool.io/lua-bundler-webpack-plugin@1.0.0 to @tarantool.io/lua-bundler-webpack-plugin@2.0.0`
+- `update multiple dependencies to latest`
 
 ## 0.0.19